test(mitm-addon): rehome usage delivery tests
